Add step specific session data identifiers
Before we had got the data for all steps and flows via 'form_data' identifier from the session. This change adapts the get_session_data and overwrite_session_data methods in such a way that they use an attribute of the steuerlotse step as the key for the data in the session. This allows us to differentiate between the different flows if we want to